requirements.txt
Flask
pysqlite3
python-dotenv>=1.0.0,<2.0.0
torch==2.4.0 
keras
PyPDF2
opencv-python
tensorflow-cpu
openai
Pillow
langchain-groq>=0.0.1,<0.1.5
playsound==1.2.2
transformers
langchain
langchain-chroma>=0.0.1,<0.1.5 
langchain-huggingface>=0.0.1,<0.1.5
langchain-text-splitters>=0.0.1,<0.1.5
langchain-community>=0.2.17,<0.3.0 